fix responsive bugs and reestructuring register and login views to login and register modal
