+++ .prettierrc
@@ -13,9 +13,7 @@
   "importOrderCaseInsensitive": true,
   "importOrderGroupNamespaceSpecifiers": true,
   "plugins": [
-    "prettier-plugin-tailwindcss",
     "@trivago/prettier-plugin-sort-imports"
-  ],
-  "tailwindConfig": "./tailwind.config.js"
+  ]
 }
