--- src/types.d.ts
@@ -1,5 +1,0 @@
-export type onProps = {
-  onclose?: () => void,
-  onShowRegister?: () => void,
-  onShowLogin?: () => void
-}
